+++ ocaml-lsp-server/test/e2e-new/code_actions.ml
@@ -1,58 +1,13 @@
 open Test.Import
+open Lsp_helpers
 
-let openDocument ~client ~uri ~source =
-  let textDocument =
-    TextDocumentItem.create ~uri ~languageId:"ocaml" ~version:0 ~text:source
+let iter_code_actions ?prep ?path ?(diagnostics = []) ~source range =
+  let makeRequest textDocument =
+    let context = CodeActionContext.create ~diagnostics () in
+    Lsp.Client_request.CodeAction
+      (CodeActionParams.create ~textDocument ~range ~context ())
   in
-  Client.notification
-    client
-    (TextDocumentDidOpen (DidOpenTextDocumentParams.create ~textDocument))
-
-let iter_code_actions ?(prep = fun _ -> Fiber.return ()) ?(path = "foo.ml")
-    ?(diagnostics = []) ~source range k =
-  let got_diagnostics = Fiber.Ivar.create () in
-  let handler =
-    Client.Handler.make
-      ~on_notification:
-        (fun _ -> function
-          | PublishDiagnostics _ -> (
-            let* diag = Fiber.Ivar.peek got_diagnostics in
-            match diag with
-            | Some _ -> Fiber.return ()
-            | None -> Fiber.Ivar.fill got_diagnostics ())
-          | _ -> Fiber.return ())
-      ()
-  in
-  Test.run ~handler @@ fun client ->
-  let run_client () =
-    let capabilities =
-      let window =
-        let showDocument =
-          ShowDocumentClientCapabilities.create ~support:true
-        in
-        WindowClientCapabilities.create ~showDocument ()
-      in
-      ClientCapabilities.create ~window ()
-    in
-    Client.start client (InitializeParams.create ~capabilities ())
-  in
-  let run =
-    let* (_ : InitializeResult.t) = Client.initialized client in
-    let uri = DocumentUri.of_path path in
-    let* () = prep client in
-    let* () = openDocument ~client ~uri ~source in
-    let+ resp =
-      let context = CodeActionContext.create ~diagnostics () in
-      let request =
-        let textDocument = TextDocumentIdentifier.create ~uri in
-        CodeActionParams.create ~textDocument ~range ~context ()
-      in
-      Client.request client (CodeAction request)
-    in
-    k resp
-  in
-  Fiber.fork_and_join_unit run_client (fun () ->
-      run >>> Fiber.Ivar.read got_diagnostics >>> Client.stop client)
+  iter_lsp_response ?prep ?path ~makeRequest ~source
 
 let print_code_actions ?(prep = fun _ -> Fiber.return ()) ?(path = "foo.ml")
     ?(filter = fun _ -> true) source range =
@@ -562,7 +517,7 @@ let f (x : t) = x
 |ocaml}
   in
   let uri = DocumentUri.of_path "foo.ml" in
-  let prep client = openDocument ~client ~uri ~source:impl_source in
+  let prep client = open_document ~client ~uri ~source:impl_source in
   let intf_source = "" in
   let range =
     let start = Position.create ~line:0 ~character:0 in